+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Find match on an expired position
+- Show loading screen when app starts with an expired position
 
 ## [1.2.6] - 2023-09-06